+++ h2buster.py
@@ -2,17 +2,45 @@ import hyper
 import threading, queue
 import ssl, sys, time, argparse
 import urllib.parse
-from socket import gaierror
+import multiprocessing
 
+# Metadata variables, program constants and global variables.
 __author__ = "https://github.com/00xc/"
-__version__ = "0.2"
-PROGRAM_INFO = "h2buster: an HTTP/2 web directory brute-force scanner."
-QUEUE_LIMIT = 100000
+__version__ = "0.3"
 
-# Global variable of extensions
+PROGRAM_INFO = "h2buster: an HTTP/2 web directory brute-force scanner."
+DASHLINE = "------------------------------------------------"
+DASHLINE_A = "=========================================="
+
+WORDLIST_MVAR = "wordlist"
+TARGET_MVAR = "target"
+DIR_DEPTH_MVAR = "directory_depth"
+CNX_MVAR = "connections"
+THREADS_MVAR = "threads"
+
+WORDLIST_DEFAULT = None
+TARGET_DEFAULT = None
+DIR_DEPTH_DEFAULT = 2
+CNX_DEFAULT = 3
+THREADS_DEFAULT = 15
+
+WORDLIST_HELP = "Directory wordlist"
+TARGET_HELP = "Target URL/IP address. Default port is 443 and HTTPS enabled. To specify otherwise, use ':port' or 'http://' (port will default to 80 then)."
+DIR_DEPTH_HELP = "Maximum recursive directory depth. Minimum is 1, default is " + str(DIR_DEPTH_DEFAULT) + "."
+CNX_HELP = "Number of HTTP/2 connections. Default is " + str(CNX_DEFAULT) + "."
+THREADS_HELP = "Number of threads per connection. Default is " + str(THREADS_DEFAULT) + "."
+
+TIMESTAMP_ROUND = 3
+TLS_DEFAULT = 1
+
 ext = ["/", "" ,".php", ".html", ".asp", ".js", ".css"]
 
-# Read "opts" options from command line
+# Function: return time since t0
+def timestamp(t0):
+	t = str(round(time.time() - t0, TIMESTAMP_ROUND))
+	return t
+
+# Function: read "opts" options from command line interface
 def read_inputs(info, opts, h, defaults, mvar):
 	parser = argparse.ArgumentParser(description=info)
 	for i, o in enumerate(opts):
@@ -22,26 +50,37 @@ def read_inputs(info, opts, h, defaults, mvar):
 	args = parser.parse_args()
 	return args
 
-# Parse target and check if its HTTPS
-def check_tls(ip):
-	ip = ip.split("://")
-	if len(ip) == 1:
-		s = 1
-		ip = ip[0]
+# Function: parse input target and decide h2 or h2c
+def parse_target(target):
+	target = target.split("://")
+	if len(target) == 1:
+		s = TLS_DEFAULT
+		url = target[0]
 	else:
-		if ip[0] == "http": s = 0
-		elif ip[0] == "https": s = 1
+		url = target[1]
+		if target[0] == "http": s = 0
+		elif target[0] == "https": s = 1
 		else: sys.exit("[-] Target not understood")
-		ip = ip[1]
-	return ip, s
 
-# Connect to target and return connection object
-def h2_connect(ip, s):
+	url = url.split("/", 1)
+	ip = url[0]
+	if len(url)==1 or len(url[1])==0:
+		start_dir = "/"
+	else: 
+		start_dir = "/" + url[1]
+		if start_dir[-1] != "/": start_dir = start_dir + "/"
+
+	return ip, start_dir, s
+
+# Function: connect via H2 to specified IP[:port]. s=1 means TLS, s=0 means h2c
+def h2_connect(s, ip):
 	# Get port
 	if len(ip.split(":")) == 1:
 		if s==1: port=443
 		elif s==0: port=80
-	else: port = int(ip.split(":")[-1])
+	else:
+		try: port = int(ip.split(":")[-1])
+		except ValueError: sys.exit("[-] Invalid URL")
 	# Start connection
 	if s == 1:
 		ctx = ssl.SSLContext()
@@ -50,136 +89,132 @@ def h2_connect(ip, s):
 		conn = hyper.HTTP20Connection(ip, port=port, ssl_context=ctx, enable_push=False)
 	elif s == 0:
 		conn = hyper.HTTP20Connection(ip, port=port, enable_push=False)
-	# Test connectivity before starting the scan
 	try: conn.connect()
-	except AssertionError: sys.exit("H2 not supported for that target.")
-	except gaierror as excp: sys.exit(excp)
-	except Exception as excp: sys.exit(excp)
-	conn.ping("00000000")
+	except AssertionError: sys.exit("[-] H2 not supported for that target.")
 	return conn, port
 
-# Thread function
-def stream_worker(q, conn, output):
+# Function: main scan function. Starts up a number of processes which handle their own h2 connection and sends them entries to scan
+def main_scan(s, ip, directory, wordlist, dir_depth, max_depth, connections, threads):
+
+	if dir_depth >= max_depth: return
 	global ext
-	results = dict()
-	redirections = dict()
-
-	try:
-		while True:
-			# Retrieve dictionary entry
-			directory, entry = q.get()
-			if entry is None: break
-			entry = entry.rstrip()
-			if entry=="":
-				q.task_done()
-				continue
-
-			# Scan wordlist entry with all extensions
-			for ex in ext:
-				sid = conn.request("HEAD", directory + urllib.parse.quote_plus(entry) + ex)
-				resp = conn.get_response(sid)
-				results[directory + entry + ex] = resp.status
-				if resp.status==301 or resp.status==302:
-					try: redirections[directory + entry + ex] = resp.headers.get(b"location")[0].decode("utf-8")
-					except TypeError: redirections[directory + entry + ex] = "NULL"
-
-			# Print results
-			for url, st in results.items():
-				if st!=404:
-					if st==301 or st==302: tail = " -> " + redirections[url]
-					else:
-						tail = ""
-						if url[-1]=="/" and st!=400: output.put(url)
-					print(url + ": " + str(st) + tail)
-
-			results.clear()
-			redirections.clear()
-			q.task_done()
-	except Exception as exc: sys.exit(exc)
-
-# Recursive threaded scan for a specific directory
-def threaded_scan(conn, directory, file, rec_level, max_rec, nthreads):
-
-	if rec_level >= max_rec: return
-
-	# Input and output queues work and results for the threads
-	work = queue.Queue(QUEUE_LIMIT)
-	output = queue.Queue(QUEUE_LIMIT)
-
-	# Start threads
-	threads = list()
-	for i in range(nthreads):
-		t = threading.Thread(target=stream_worker, args=(work, conn, output))
-		t.daemon = True
-		t.start()
-		threads.append(t)
-		time.sleep(0.05)
 
 	print("\n[*] Starting scan on " + directory)
 
-	# Put entries to be scanned into work queue
-	with open(file, "r") as f:
-		for entry in f:
-			work.put((directory, entry))
+	# Start input/output queues
+	manager = multiprocessing.Manager()
+	output = manager.list()
+	inwork = manager.Queue(connections)
 
-	# Block until the queue is over
-	work.join()
+	# Start connections
+	pool = list()
+	for i in range(connections):
+		p = multiprocessing.Process(target=process_worker, args=(s, ip, threads, inwork, output))
+		p.daemon = True
+		p.start()
+		pool.append(p)
 
-	# Send stop signal to threads
-	for i in range(nthreads):
-		work.put(("", None))
-	for t in threads:
-		t.join()
+	# Put work
+	c = 0
+	try:
+		with open(wordlist, "r") as f:
+			for entry in f:
+				entry = entry.rstrip()
+				for ex in ext:
+					c+=1
+					inwork.put((directory, entry+ex))
+	except FileNotFoundError:
+		for i in range(connections*threads):
+			inwork.put((None, None))
+		for p in pool: p.join()
+		sys.exit("[-] Wordlist file not found.")
 
-	# Call recursively with each found directory
-	while not output.empty():
-		fd = output.get()
-		threaded_scan(conn, fd, file, rec_level+1, max_rec, nthreads)
+	# Send kill signals
+	for i in range(connections*threads):
+		inwork.put((None, None))
 
+	# Block until processes are done
+	for p in pool: p.join()
+
+	# Recursive calls with found directories
+	for d in output:
+		main_scan(s, ip, d, wordlist, dir_depth+1, max_depth, connections, threads)
+
+# Function: process worker. Starts a connection and a number of threads that perform requests on that connection.
+def process_worker(s, ip, threads, inwork, output):
+	conn, port = h2_connect(s, ip)
+
+	threadpool = list()
+	for i in range(threads):
+		t = threading.Thread(target=thread_worker, args=(conn, inwork, output))
+		t.daemon = True
+		t.start()
+		threadpool.append(t)
+
+	for t in threadpool: t.join()
+	conn.close()
+
+# Function: thread worker. For each entry in the inwork queue, sends one request and reads response status code
+def thread_worker(conn, inwork, output):
+	while True:
+		directory, entry = inwork.get()
+		if entry is None: break
+		
+		sid = conn.request("HEAD", directory + entry.replace(" ", "%20"))
+		resp = conn.get_response(sid)
+
+		st = resp.status
+		headers = resp.headers
+
+		if st!=404:
+			if st==301 or st==302: tail = " -> " + headers.get(b"location")[0].decode("utf-8")
+			else:
+				tail = ""
+				if st==200 and entry[-1]=="/": output.append(directory + entry)
+			print(directory+entry + ": " + str(st) + tail)
+
+# Main start point. Read, verify inputs and call main_scan()
 if __name__ == '__main__':
 
-	print("--------------------------------")
+	print(DASHLINE)
 	print("h2buster v" + __version__)
-	print("--------------------------------")
+	print(DASHLINE)
 
-	# Input reading
-	opts = ["w", "u", "r", "t"]
-	mvar = ["wordlist", "target", "directory_depth", "threads"]
-	h = ["Directory wordlist", "Target URL/IP address. Default port is 443 with HTTPS. To specify otherwise, use ':port' or 'http://' (port will default to 80 then).", "Maximum directory depth. Minimum is 1, default is 2.", "Number of threads. Default is 10."]
-	defaults = [None, None, 2, 10]  # None => argument is required
+	# Read CLI inputs
+	opts = ["w", "u", "r", "c", "t"]
+	mvar = [WORDLIST_MVAR, TARGET_MVAR, DIR_DEPTH_MVAR, CNX_MVAR, THREADS_MVAR]
+	h = [WORDLIST_HELP, TARGET_HELP, DIR_DEPTH_HELP, CNX_HELP, THREADS_HELP]
+	defaults = [WORDLIST_DEFAULT, TARGET_DEFAULT, DIR_DEPTH_DEFAULT, CNX_DEFAULT, THREADS_DEFAULT]
 	args = read_inputs(PROGRAM_INFO, opts, h, defaults, mvar)
 
+	# Input checking
 	try:
-		# Input checking
-		try:
-			args.r = int(args.r)
-			args.t = int(args.t)
-			if args.t<1 or args.r<1:
-				sys.exit("\n[-] threads and recursion_depth must be greater than zero.")
-		except ValueError:
-			sys.exit("\n[-] Invalid non-numerical option introduced.")
+		args.r = int(args.r)
+		args.c = int(args.c)
+		args.t = int(args.t)
+		if args.t<1 or args.r<1:
+			sys.exit("[-] " + THREADS_MVAR + " and " + DIR_DEPTH_MVAR + " must be greater than zero")
+	except ValueError:
+		sys.exit("[-] Invalid non-numerical option introduced")
 
-		# For benchmarking purposes
-		t0 = time.time()
+	# Start timer (for benchmarking purposes)
+	t0 = time.time()
 
-		# Check wheter we should use HTTP or HTTPS and start connection
-		ip, s = check_tls(args.u)
-		conn, port = h2_connect(ip, s)
+	# Parse target URL
+	ip, start_dir, s = parse_target(args.u)
 
-		print("[+] Connected to " + ip + " on port " + str(port))
-		if s==1: print("[*] TLS is ON")
-		else: print("[*] TLS is OFF")
-		print("[*] Number of threads: " + str(args.t))
-		print("[*] Directory depth: " + str(args.r))
+	# Print info
+	print("[*] Initializing scan on " + ip)
+	if s==1: print("[*] TLS is ON")
+	else: print("[*] TLS is OFF\n")
+	print("[*] Base directory: " + start_dir)
+	print("[*] Maximum directory depth: " + str(args.r) + " (base directory is depth 1)")
+	print("[*] Checking for file extensions: " + ", ".join(["'"+ex+"'" for ex in ext]))
+	print("[*] Number of connections: " + str(args.c))
+	print("[*] Number of threads per connection: " + str(args.t))
+	print(DASHLINE)
 
-		# Start threaded scan. This will call itself recursively with found directories args.r-1 times
-		try: threaded_scan(conn, "/", args.w, 0, args.r, args.t)
-		except Exception as exc: sys.exit(exc)
+	# Start main scan which will call itself for each found directory
+	main_scan(s, ip, start_dir, args.w, 0,  args.r, args.c, args.t)
 
-		conn.close()
-
-		print(" \n[*] Program ran in " + str(round(time.time()-t0, 3)) + " seconds.")
-
-	except KeyboardInterrupt:
-		conn.close()
-		print("\r[-] Scan aborted after " + str(round(time.time()-t0, 3)) + " seconds.")
+	print("\n[*] Program ran in " + timestamp(t0) + " seconds")